refactor(optim): rename sgd module -> optim
