+++ lfa/src/approximators/pair.rs
@@ -33,7 +33,7 @@ impl Approximator for PairFunction {
         ])
     }
 
-    fn update_with<O: crate::sgd::Optimiser>(&mut self, opt: &mut O, f: &Features, es: [f64; 2]) -> UpdateResult<()> {
+    fn update_with<O: crate::optim::Optimiser>(&mut self, opt: &mut O, f: &Features, es: [f64; 2]) -> UpdateResult<()> {
         opt.step(&mut self.weights.column_mut(0), f, es[0])
             .and(opt.step(&mut self.weights.column_mut(1), f, es[1]))
     }
@@ -46,7 +46,7 @@ mod tests {
     use crate::{
         Approximator,
         basis::{Projector, Fourier, TileCoding},
-        sgd::SGD,
+        optim::SGD,
     };
     use std::hash::BuildHasherDefault;
     use super::PairFunction;
