+++ lfa/src/approximators/vector.rs
@@ -32,7 +32,7 @@ impl Approximator for VectorFunction {
         Ok(f.matmul(&self.weights.view()).into_raw_vec())
     }
 
-    fn update_with<O: crate::sgd::Optimiser>(&mut self, opt: &mut O, f: &Features, es: Vec<f64>) -> UpdateResult<()> {
+    fn update_with<O: crate::optim::Optimiser>(&mut self, opt: &mut O, f: &Features, es: Vec<f64>) -> UpdateResult<()> {
         es.into_iter()
             .zip(self.weights.gencolumns_mut().into_iter())
             .fold(Ok(()), |acc, (e, mut c)| acc.and(opt.step(&mut c, f, e)))
@@ -46,7 +46,7 @@ mod tests {
     use crate::{
         Approximator,
         basis::{Projector, Fourier, TileCoding},
-        sgd::SGD,
+        optim::SGD,
     };
     use std::hash::BuildHasherDefault;
     use super::VectorFunction;
